Window Proc Time with time boundaries

+++ flink-libraries/flink-table/src/main/java/org/apache/flink/table/plan/nodes/datastream/function/DataStreamProcTimeWindowAggregator.java
@@ -97,7 +97,7 @@ public abstract class DataStreamProcTimeWindowAggregator implements Serializable
 
 	private void setCountAggregation(int i) {
 
-		if (typeOutput.get(i).getTypeClass().equals(Integer.class)) {
+		if (typeOutput.get(i).getTypeClass().equals(Long.class)) {
 			aggregatorImpl.add(new AnyCounterAggregator());
 		} else {
 			throw new IllegalArgumentException("Unsupported aggregation type");
